Generic test runner: Migrate HTML Checkbox (13/n)
[ghstack-poisoned]
