+++ packages/component-driver-html-test/src/examples/index.ts
@@ -1,4 +1,4 @@
-export * from './HTMLCheckbox.examples';
+export * from './checkbox';
 export * from './HTMLRadioButtonGroup.examples';
 export * from './HTMLSelect.examples';
 export * from './input';
